// src/app/model/organization.ts
export class Organization {
  id: number = 0;
  name: string = "";
  contactPerson: string = "";
  contactNumber: string = "";
  supportNumber: string = "";
  supportEmail: string = "";
  addressLine1: string = "";
  addressLine2: string = "";
  addressLine3: string = "";
  stateId: number = 0;
  cityId: number = 0;
  pinCode: string = "";
  countryId: number = 0;
}